+++ src/main/java/inflearn/springadvance/proxy/config/v5_autoproxy/AutoProxyConfig.java
@@ -6,6 +6,7 @@ import inflearn.springadvance.proxy.config.v3_proxyfactory.advice.LogTraceAdvice
 import inflearn.springadvance.proxy.trace.logtrace.LogTrace;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.aop.Advisor;
+import org.springframework.aop.aspectj.AspectJExpressionPointcut;
 import org.springframework.aop.support.DefaultPointcutAdvisor;
 import org.springframework.aop.support.NameMatchMethodPointcut;
 import org.springframework.context.annotation.Bean;
@@ -18,11 +19,22 @@ import org.springframework.context.annotation.Import;
 
 public class AutoProxyConfig {
 
-    @Bean
+//    @Bean
     public Advisor advisor1(LogTrace logTrace) {
         //pointcut
         NameMatchMethodPointcut pointcut = new NameMatchMethodPointcut();
         pointcut.setMappedNames("request*", "order*", "save*");
+
+        //advice
+        LogTraceAdvice advice = new LogTraceAdvice(logTrace);
+        return new DefaultPointcutAdvisor(pointcut, advice);
+    }
+
+    @Bean
+    public Advisor advisor2(LogTrace logTrace) {
+        //pointcut
+        AspectJExpressionPointcut pointcut = new AspectJExpressionPointcut();
+        pointcut.setExpression("execution(* inflearn.springadvance.proxy.app..*(..))");
 
         //advice
         LogTraceAdvice advice = new LogTraceAdvice(logTrace);